+++ DTOs/Requests/ExamScheduleRequest.cs
@@ -7,9 +7,9 @@
         public string? Type { get; set; }
         public bool? Form { get; set; }
         public string? Status { get; set; }
-        public long AcademicYearId { get; set; }
-        public long Subject { get; set; }
-        public long SemesterId { get; set; }
-        public long GradeLevelsId { get; set; }
+        public int AcademicYearId { get; set; }
+        public int Subject { get; set; }
+        public int SemesterId { get; set; }
+        public int GradeLevelsId { get; set; }
     }
 }
