+++ DTOs/Requests/SubjectGroupRequest.cs
@@ -8,6 +8,6 @@ namespace ISC_ELIB_SERVER.DTOs.Requests
         [MaxLength(100, ErrorMessage = "Tên không được vượt quá 100 ký tự")]
         public string Name { get; set; }
         [Required(ErrorMessage = "Id teacher không được để trống")]
-        public long TeacherId { get; set; }
+        public int TeacherId { get; set; }
     }
 }
