+++ DTOs/Requests/TeacherInfoRequest.cs
@@ -2,7 +2,7 @@
 {
     public class TeacherInfoRequest
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string? Cccd { get; set; }
         public DateTime? IssuedDate { get; set; }
         public string? IssuedPlace { get; set; }
@@ -11,11 +11,11 @@
         public string? UnionPlace { get; set; }
         public bool? PartyMember { get; set; }
         public DateTime? PartyDate { get; set; }
-        public long? UserId { get; set; }
+        public int? UserId { get; set; }
         public string? AddressFull { get; set; }
-        public long ProvinceCode { get; set; }
-        public long DistrictCode { get; set; }
-        public long WardCode { get; set; }
+        public int ProvinceCode { get; set; }
+        public int DistrictCode { get; set; }
+        public int WardCode { get; set; }
 
     }
 }
