+++ DTOs/Requests/TestQuestionRequest.cs
@@ -5,10 +5,10 @@ namespace ISC_ELIB_SERVER.DTOs.Requests
     public class TestQuestionRequest
     {
 
-        public long? TestId { get; set; }
+        public int? TestId { get; set; }
         public string? QuestionText { get; set; }
         public string? QuestionType { get; set; }
 
-       
+
     }
 }
