+++ DTOs/Requests/TestSubmissionAnswerRequest.cs
@@ -5,13 +5,13 @@ namespace ISC_ELIB_SERVER.DTOs.Requests
     public class TestSubmissionAnswerRequest
     {
         [Required(ErrorMessage = "SubmissionId is required.")]
-        public long SubmissionId { get; set; }
+        public int SubmissionId { get; set; }
 
         [Required(ErrorMessage = "QuestionId is required.")]
-        public long QuestionId { get; set; }
+        public int QuestionId { get; set; }
 
         [Required(ErrorMessage = "SelectedAnswerId is required.")]
-        public long SelectedAnswerId { get; set; }
+        public int SelectedAnswerId { get; set; }
 
         public string? AnswerText { get; set; }
     }
