+++ DTOs/Responses/CampusResponse.cs
@@ -4,11 +4,11 @@ namespace ISC_ELIB_SERVER.DTOs.Responses
 {
     public class CampusResponse
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string? Name { get; set; }
         public string? Address { get; set; }
         public string? PhoneNumber { get; set; }
-        public long SchoolId { get; set; }
-        public long UserId { get; set; }
+        public int SchoolId { get; set; }
+        public int UserId { get; set; }
     }
 }
