+++ DTOs/Responses/ChatResponse.cs
@@ -2,10 +2,10 @@
 {
     public class ChatResponse
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Content { get; set; } = string.Empty;
-        public long UserId { get; set; }
-        public long SessionId { get; set; }
+        public int UserId { get; set; }
+        public int SessionId { get; set; }
         public DateTime? SentAt { get; set; }
     }
 }
