+++ DTOs/Responses/ClassesResponse.cs
@@ -4,7 +4,7 @@ namespace ISC_ELIB_SERVER.DTOs.Responses
 {
     public class ClassesResponse
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public string? Code { get; set; }
 
@@ -17,13 +17,13 @@ namespace ISC_ELIB_SERVER.DTOs.Responses
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
         public string? Description { get; set; }
 
-        public long GradeLevelId { get; set; }
+        public int GradeLevelId { get; set; }
 
-        public long AcademicYearId { get; set; }
+        public int AcademicYearId { get; set; }
 
-        public long UserId { get; set; }
+        public int UserId { get; set; }
 
-        public long ClassTypeId { get; set; }
+        public int ClassTypeId { get; set; }
 
 
     }
