+++ DTOs/Responses/NotificationResponse.cs
@@ -2,11 +2,11 @@
 {
     public class NotificationResponse
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Title { get; set; }
         public string Content { get; set; }
         public string Type { get; set; }
-        public long SenderId { get; set; }
-        public long UserId { get; set; }
+        public int SenderId { get; set; }
+        public int UserId { get; set; }
     }
 }
