+++ DTOs/Responses/SubjectGroupResponse.cs
@@ -4,8 +4,8 @@ namespace ISC_ELIB_SERVER.DTOs.Responses
 {
     public class SubjectGroupResponse
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string? Name { get; set; }
-        public long TeacherId { get; set; }
+        public int TeacherId { get; set; }
     }
 }
